broke reward into terms and messed with action scaling
there is a problem with the scaling on the action space. the action that
is being passed to the ``step()`` method is not within the bounds. this
is from the SB3 side of things. also, when wrapping the ``env()`` in a
``RescaleAction()`` wrapper, this is seemingly not applied.

broke reward into terms. need to work on adding an angular reward term
for aligning the axis of the agent with the goal.