fixed the angular reward with the help of Minh
had to include the |diff - pi| to get it to correctly wrap around. need
to do extensive reward engineering at this point. as per a discussion
Minh and I had, it is believed that the inclusion of positive and
negative reward streams is complicating the agent's ability to learn.

need to fix this