made changes to nav2d.py, to implement the reward changes
*added the maximum distance to the class, used in
``_set_observation_space()`` and ``step()`` methods.
*changed absolute difference in heading such that it aligns with a
triangular function that is ``0`` at angles of ``0`` and ``2pi``, and decreases linearly to
``-1`` at an angle of ``pi``.
*added several reward terms that have magnitudes at or around ``-1``.
focused primarly on negative reward terms within the range ``[-1,0]``,
with the only exception being the ``rew_approach`` term, which is
positive when heading toward the goal.
*reward now consists of a heading term, a time term, another alignment
term, and an approach term.
*also did minor commenting here and there.