added vectorization flags, function for evaluation, and model saving
added a vectorization flag that creates the environment based on whether
this flag is active. if ``vectorize``, creates a vectorized environment,
and a regular environment if else. also, an eval environment is created
for evaluation.

added a spot for running several trials and saving the model
incrementally at different points in the execution of this cell. this
will allow for testing the model at different points in its training.