add argument to specify whether a Nav2D env is for evaluation
Evaluation environments are always randomized at every reset to test the generalizability of a trained model to different evaluation scenarios. The randomization bounds should also be the maximum possible amount.

The purpose of this is such that one can evaluate the true performance of a model and report the mean evaluation reward as a future Optuna objective function