made several changes to 'nav2d.py', including to the reward
changed the frequency at which the agent position is randomized to 5
from 1.

fixed concatenation in ``_get_obs()`` method, from 'np.concat' to
'np.concatenation'.

fixed heading calculation to use the angle after stepping the model, so
that it is reflective of the current model state.

changed reward to use a positive heading reward, an approach reward, and
a lessened time reward.